refactor: classify pagination, filter dispatch
